// src/api/generated/models/listInfosApiV1InfosGetParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */
import type { InfoCategory } from './infoCategory'
import type { InfoSortField } from './infoSortField'
import type { InfoStatus } from './infoStatus'
import type { SortOrder } from './sortOrder'

export type ListInfosApiV1InfosGetParams = {
  office_id?: string | null
  category?: InfoCategory | null
  status?: InfoStatus | null
  starts_from?: string | null
  ends_to?: string | null
  sort_by?: InfoSortField
  order?: SortOrder
  /**
   * Bounding Box: minLon,minLat,maxLon,maxLat
   */
  bbox?: string | null
  /**
   * @minimum 1
   */
  page?: number
  /**
   * @minimum 1
   * @maximum 100
   */
  size?: number
  q?: string | null
}